--- addons/example_addon/libraries.txt
@@ -1,1 +1,0 @@
-express=5.2.1
